backend/requirements_backup.txt
# FastAPI Backend Dependencies for AWS Deployment

# Web Framework
FastAPI==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1

# Data Processing
pandas==2.1.3
xlsxwriter==3.1.9
openpyxl==3.1.2

# Database
pymysql==1.1.0
psycopg2-binary==2.9.9

# System Monitoring
psutil==5.9.6

# Utilities
python-socketio==5.9.0
eventlet==0.33.3

# Development
python-dotenv==1.0.0

# Supabase
supabase==2.0.2

# Additional AWS dependencies
boto3==1.34.0
requests==2.31.0 